% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInit.R
\name{BulkDataInit}
\alias{BulkDataInit}
\title{Title}
\usage{
BulkDataInit(exprFile, sampleFile, CellType = "HCC",
  Genelistfile = "none", IfSaveFile = F, SaveFileDir = "none")
}
\arguments{
\item{SaveFileDir}{}
}
\description{
Title
}
