% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionCorrNet.R
\name{ConditionCorrNet}
\alias{ConditionCorrNet}
\title{Title}
\usage{
ConditionCorrNet(exprMatrix, geneList, sampleNum = 10,
  sampleSize = 100, disMethod = "spearman", pairNum = 10000,
  orderScore = "dis", IfSaveFile = F, saveFile = "none")
}
\arguments{
\item{saveFile}{}
}
\description{
Title
}
