% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInit.R
\name{TenxDataInit}
\alias{TenxDataInit}
\title{Title}
\usage{
TenxDataInit(exprMatrix, Genelistfile = "none", IfSaveFile = F,
  SaveFileDir = "none")
}
\arguments{
\item{SaveFileDir}{}
}
\description{
Title
}
\examples{

}
